+++ RunningGo.API/SistemaDeMetas/Resources/ProcessResource.cs
@@ -8,6 +8,6 @@ public class ProcessResource
     public DateTime Date { set; get; }
     
     //Relationships
-    public int UserId { set; get; }
+    public long UserId { set; get; }
     public IList<GoalResource> Goals { set; get; } = new List<GoalResource>();
 }
